requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
cffi==1.17.0
channels==4.1.0
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.0
daphne==4.1.2
Django==5.1
django-formtools==2.5.1
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
greenlet==3.0.3
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
zope.interface==7.0.1